[video][ios] Load track information in asynchronous context (#40355)
# Why

Event handlers are most often run on the main queue (property changes we
listen to happen on the main queue) and we are using some deprecated
synchronous methods to load track information when a change is detected,
this can cause main thread lockups for slow-loading sources.
New methods, which load the track information asynchronously are
available, but quite a lot of changes is needed to adapt them to the
current, synchronous nature of expo-video event system.

# How

Added a TaskManager actor to `VideoPlayerObserver`. When the observer
has to do asynchronous loading of track data, the load should go through
the task manager. The manager will cancel the current load (if exists)
and add the current load into the tasks dictionary.

# Test Plan

Tested in BareExpo on iOS