[UI][iOS] Enhance PaddingModifier to support default SwiftUI padding when no custom values are provided (#40715)
# Why

Currently, the padding modifier requires passing an object defining the
desired padding. But SwiftUI padding modifier comes with a system
default padding that is applied when no value is provided.

This pull request makes the parameters to the padding modifier optional
and if they are not provided it defaults to the system padding.

This PR also introduces a small styling fix when using labels without an
image. Previously we were adding an empty string if the image was not
provided, which was causing an empty space. Now we just check if the
image is not provided, then we just not render anything.

# How

Validate if the padding has custom parameters, if so we keep the current
implementation otherwise we use the default padding.

For `Label.swift` added a validation for system image and
`.labelStyle(.titleAndIcon)` which ensures proper layout when icon
exists

# Test Plan

<img width="426" height="432" alt="image"
src="https://github.com/user-attachments/assets/bb82b0de-c8f2-4b66-bec3-4758b18a43b7"
/>

# Checklist

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [x] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [x] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)