[video][android] Fix `player` prop not working when re-assigned a previous player. (#40709)
# Why

Fixes https://github.com/expo/expo/issues/40512

When doing player assignment loop to the same view like playerA ->
playerB -> playerA.
The last assignment will do nothing, this is because when we switch to
playerB, we don't inform the playerA that it's view has been changed.
Because of this upon re-assignment (playerB -> playerA) nothing happens,
as playerA thinks it's already mounted in the video view.

# How

When changing the player of a video view call `player.
changePlayerView(null)` on the old player to let it know it's not
connected to a view.

# Test Plan

Tested in BareExpo on Android

(cherry picked from commit 197fbc12a43b2e7a76e0d48565dd4a8602c310d4)