[go][quest] Add auth workaround for Meta Quest (#40706)
# Why

We use `expo-web-browser` custom tabs for auth in Expo Go. Redirects
back into the app seem to be broken in custom tabs on quest, which
prevents login.

It's likely a regression, before the login used to work.

# How

As a workaround I have created a modified quest-specific version of the
auth polyfill that uses a regular browser instead of custom tabs. This
makes the login work but has the downside of the browser not closing
after the login is fulfilled.

This is not ideal, but until we get a solution from Meta it's probably
the best we can do.

# Test Plan

Tested in Expo Go on the SDK-54 branch (main is not building for me)