fix(log-box): empty HMR update should not reset the error overlay (#40741)
The empty updates are ignored by the client and don't change the runtime
in any way, the error overlay should not be reset, as it will cause
errors from non empty updates to disappear.

In the past the empty updates were caused by updating project file that
are not part of the bundle programmatically by the CLI, I don't see that
happening at the moment, but still some empty updates are sent.

Related PR from the past
- https://github.com/expo/expo/pull/39835