breaking(cli): Drop `EXPO_USE_FAST_RESOLVER` flag and implementation (#40673)
Resolve DVT-8

## Summary

We're not planning to maintain two different resolvers, since the
`package.json` fields for CommonJS or mixed resolution, and `exports`
resolution can introduce subtly different changes which are hard to
track, and can cause regressions. Metro's resolution has since become
faster and the original purpose of the fast resolver is partially
obsolete. If we're finding that Metro resolution is too slow there's
other things we can do or improve.

If we find that we need to fork anything, the future plan is to instead
fully fork, without introducing two complicated code paths that need
testing, since it's easier to maintain fewer modes and combinations of
Metro customisations.

We've previously needed the fast resolver for other flags, but we've
disabled the auto-flipping of the `EXPO_USE_FAST_RESOLVER` flag in
#39073 and this has been holding up fine. For SDK 55, we'd like to
remove the fast resolver code path entirely.

## Set of changes

- Drop fast resolver (remove `createExpoMetroResolver.ts`,
`createJResolver.ts`, `formatFileCandidates.ts`)
- Replace `FailedToResolvePathError` for native-only error with custom
error
- Drop `isFastResolverEnabled` / `EXPO_USE_FAST_RESOLVER` entirely
- Remove direct dependencies on `resolve` and `resolve.exports`

An unrelated change in this PR is the removal of `resolve` in
`createServerRouteMiddleware.ts`. The resolution there of `route.file`
(from the generated manifest output) was redundant since `file: string`
already points to the exact relative file path.

## Test plan

The existing tests should pass without further changes. As of opening
this PR, Windows CI is failing due to unrelated changes in #40594. This
doesn't block this PR however, since they're Windows only and the Linux
E2E CI workflows are passing.