[ios][expo-updates-interface] New interface for metrics (#40805)
# Why

Expose a limited set of data on the currently running update, etc., to native code, without requiring an `expo-updates` dependency.

# How

Instead of using the existing protocol (designed for the needs of `expo-dev-launcher`), add a new smaller protocol specifically for this use case. Only the enabled updates controller implements this protocol.

# Test Plan

- CI should pass
- Use a test app to verify that updates data can be read through the interface