[video][android] Fix surface view not showing when changing the `player` prop (#40817)
# Why

Fixes https://github.com/expo/expo/issues/38426

On Android there is an issue with the surface view getting hidden for
already playing players when they are placed into a fresh `VideoView`

# How

This happens because when the player is already playing the
`onFirstFrameRendered` event is not emitted, therefore we don't update
the visibility.

Check if the player has already rendered a frame. In that case we can
just display it in the surface view.

# Test Plan

Tested in BareExpo on Android and iOS on a new testing screen (will add
it in a separate PR since I'll be backporting this one to SDK 54)