[android][expo-updates-interface] New interface for metrics (#40873)
# Why

Expose a limited set of data on the currently running update, etc., to native code, without requiring an `expo-updates` dependency.

(These changes correspond to similar changes merged for iOS in #40805 )

# How

Instead of using the existing interface (designed for the needs of `expo-dev-launcher`), add a new smaller interface specifically for this use case. Only the enabled updates controller implements this interface.

# Test Plan

- CI should pass
- Use a test app to verify that updates data can be read through the interface