fix(babel-preset-expo): Readd `caller.isHMREnabled` check and add `enableReactFastRefresh` option (#40774)
Partial regression from #39826
Resolves #40569
Resolves DVT-54

We have a couple of use-cases for `babel-preset-expo` that run the Babel
preset on environments that may flip/activate the `isDev` environment
without supporting Fast Refresh.

We should readd the `isHMREnabled` flag check (which is a flag that in
`@expo/metro-config` is now always `true` (but not set e.g. in
`jest-expo`), while also adding an `enableReactFastRefresh` option to
force-disable fast refresh if that's required, to clarify that it's an
optional addition that's enabled by default

- Update tests to set `isHMREnabled: true` on callers to reflect
`@expo/metro-config` better
- NOTE: The shared type isn't properly exposed and we should do better
in making it clear which parts of it are optional (`projectRoot` is
required when it shouldn't be). That's out of scope for this PR
- Readd `isHMREnabled` check for fast refresh
- Allow `react-refresh/babel` to be force-disabled with
`enableReactFastRefresh: false`
- Allow `react-refresh/babel` to be force-enabled with
`enableReactFastRefresh: true`

- Unit tests should cover this as-is

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)