Update Elixir SDK link in documentation
The prior exponent server SDK for elixir works fine, but it hasn't been updated in 7 years. I updated the project to remove a bunch of compile time warnings I've been ignoring and opened a pull request, but it doesn't look like the author is active on GitHub anymore. In lieu of merging my changes to their repo, I figured maybe I should ask you all to update the link in your docs to version with more modern dependencies and fewer compile time warnings.

I'm not sure if this is the best way to handle the situation. At least here we'll be able to have a conversation about what we should do, even if this isn't quite right.