[video][android] Emit `onFirstFrameRender` when player is moved to a new `VideoView` (#40854)
# Why

Currently we only emit `onFirstFrameRender` only on the first frame of
the current media source of the player.
If the player has already rendered the first frame and then was moved
into a new view, it will not emit the event. This is incorrect as we'd
like to emit the event every time the first frame is displayed in the
`VideoView`

# How

Update our first frame event emission filter to take into account
`VideoView` changes. This required quite a lot of changes, since we now
need to keep a closer track of VideoView changes on the player side.

# Test Plan

Tested in BareExpo on Android