[video] Allow assigning `null` value to the `player` prop of the `VideoView` (#40860)
# Why

Some users may choose to display an empty `VideoView` until they are
ready for playback.
On Android and iOS we already handle the case of the video player being
null, web requires just a few changes, so it's easy to add this feature.

# How

Allow for null player prop.

# Test Plan

Tested in BareExpo on Android, iOS and Web