[android][video] Fix remote media controls (#40728)
# Why

fixes https://github.com/expo/expo/issues/39450

With changes to `ExpoVideoPlaybackService` registration introduced in
SDK 54, the media controls (e.g. bluetooth play/pause) have stopped
working when the player is not bound to the service.

# How

Added a basic MediaSession to every player so that it handles media
controls events.

# Test Plan

Tested in BareExpo on Android.