[ci] increased disk size for Android emulator (#41221)
# Why

When running the Android e2e tests we often get the following errors and
a failing test:

Not enough space to create userdata partition. Available: 5303.36 MB at
/home/runner/.android/avd/avd-36.avd, need 7372.80 MB.

# How

This commit tries to fix this by setting the disk-size parameter in the
emulator setup action. The disk-size is set to 4GB which should be
enough for Android, BareExpo and Maestro.

# Test Plan

✅ Test-run:
https://github.com/expo/expo/actions/runs/19671194383/job/56343245842 -
Green CI