[ncl] Add a screen for testing mounting `VideoPlayers` in `VideoViews` (#40827)
# Why

I've recently fixed two pretty critical issues related to changing the
`VideoView` a `VideoPlayer` is currently playing in. I believe we can
prevent this in a future by having a screen for testing a scenario where
there are two players and they are moved between `VideoViews`

# How

Added a screen where there are two players, which can be advanced
between four views. There are also two buttons for simulating two
players competing for a single video view.

# Test Plan

Tested in BareExpo on Android and iOS