[video][web] Fix video pausing when entering fullscreen in electron apps (#40989)
# Why

Fixes https://github.com/expo/expo/issues/40743

In electron apps the video pauses when entering the fullscreen while
muted.

# How

Electron pauses the video when entering fullscreen, and unpauses it
after leaving fullscreen. Because of our play/pause handling the unpause
is interfered with. The fix is to only pause mounted players that are
actually playing.

See https://github.com/expo/expo/issues/40743#issuecomment-3472643142
for more details.

# Test Plan

Tested in ncl on web and also in the provided electron app repro.