[android] Add `android:maxSdkVersion` annotation to `READ_EXTERNAL_STORAGE` (#40976)
# Why

https://github.com/expo/expo/pull/40973 adds checks for conflicting
`android:maxSdkVersion` declarations.

The new plugin found issues with `READ_EXTERNAL_STORAGE`.

`expo-image` already declares the permission with
`android:maxSdkVersion` of 32, while in other places we declare it
without the annotation.

The `WRITE_EXTERNAL_STORAGE` and `READ_EXTERNAL_STORAGE` do nothing
above that target, so we can safely limit it in all places. This will
not have any real effect on peoples projects, except for limiting the
output from the gradle plugin.

# How

Add annotations in other places

# Test Plan

Tested in BareExpo on Android