[ios][precompiled] added expomodulejsi to ignore force-static-linking list
When using `USE_FRAMEWORKS` on iOS with precompiled frameworks, we have a list of expo modules that should be ignored since when they're using the React framework they fail since they expose headers from the React framework.

After splitting out the iOS JSI layer in #40755 we got another POD target that needs to be on the ignore list.

This commit fixes this by adding the ExpoModulesJSI to the ignore list.