Cast to JSONSerializable in CDP MessageHandler
JSONSerializable is defined using readonlys. The way in which the types
infer doesn't allow us to define this as a base-type without writing a
complex inference pattern, even as a generic `extends`.

Instead, we can assume that we're only going to pass serializable data
by default and loosen the type, casting as needed.