fix(expo-router): Replace mock `new URL` parsing base URL with `file:` (#41338)
# Why

We use a dummy/mock URL in the `new URL` call in `getStateFromPath` to
parse the URL while being aware of encoding. However, this leaves a mock
URL in our sourcecode that'll also be present in output bytecode.

Code scanners may pick this URL up, and flag it as unrelated to the app,
since it's a mock URL. While the `.example` domain should make it clear
that it's not a real URL, we can avoid this instead.

The WhatWG sections for URL parsing seem to indicate that this is fine.
It's hard to untangle though, so I'm not 100% certain. Parsing behaviour
seems to however be identical for `file:`:
- https://url.spec.whatwg.org/#scheme-state
- https://url.spec.whatwg.org/#file-state

# How

- Switch to a `file:` base URL for `new URL` parsing
- Hide away `url` on return value and instead only expose `hash`
- Simplify return path

# Test Plan

- Existing tests pass

# Checklist

<!--
Please check the appropriate items below if they apply to your diff.
-->

- [x] I added a `changelog.md` entry and rebuilt the package sources
according to [this short
guide](https://github.com/expo/expo/blob/main/CONTRIBUTING.md#-before-submitting)
- [ ] This diff will work correctly for `npx expo prebuild` & EAS Build
(eg: updated a module plugin).
- [ ] Conforms with the [Documentation Writing Style
Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)