[core][ios] Revert AppContext in SwiftUI views environment (#41112)
# Why

Unfortunately, changes introduced in
https://github.com/expo/expo/pull/39207 caused a memory leak of
`appContext` when app is being reloaded.

# How

As a workaround, I found a solution to wrap the value with a struct that
holds a weak reference to `AppContext`.

```swift
extension ExpoSwiftUI {
  public struct WeakAppContext: @unchecked Sendable {
    weak var value: AppContext?
  }

  public struct AppContextKey: EnvironmentKey {
    public static let defaultValue: WeakAppContext = WeakAppContext()
  }
}

extension EnvironmentValues {
  public var appContextHolder: ExpoSwiftUI.WeakAppContext {
    get { self[ExpoSwiftUI.AppContextKey.self] }
    set { self[ExpoSwiftUI.AppContextKey.self] = newValue }
  }
}
```

With this change, the API would look like this:

```swift
@Environment(\.appContext) private var appContext;

appContext.value
```

However, since this isn't currently being used anywhere, and we have an
alternative (`props.appContext`), I believe we can safely remove it for
now.

> @Kudo https://github.com/expo/expo/pull/39207#issuecomment-3369069806
> fwiw, expo-ui having the `props.appContext` since #39231 and uses it
across all views to support modifiers. we don't use app context from
environment


# Test Plan

* Open a screen with SwiftUI views like `expo-maps` or `@expo/ui`
* Reload the app
* Check if `deinit` in `AppContext` was called