chore(cli): Clean up typings and tighten up strictness in `@expo/cli` (#41335)
#41142 (metro upgrade) taskes precedence over this PR.

This is meant to tighten up type strictness, to prevent us from making
mistakes while moving around Metro patches and other code. It basically
removes all remaining `@ts-expect-error` comments in `src`. This is
after we discovered in a past Metro upgrade that a few can accidentally
hide a few type issues and mismatches. A few removals of
`ts-expect-error` in this PR already flag a few types issues.

The most significant change might be the removal of the
`runtimeBytecodeVersion` URL param flag, which did occur in a few test
fixtures. This flag looks to be unused and should be removed. But this
might now cause test mismatches.