feat(Expo Go): migrate more iOS to PBXFileSystemSynchronizedRootGroup (#41467)
# Why

- Related to https://github.com/expo/expo/pull/41464
- Migrate the majority of the Expo Go iOS codebase (minus tests since
it's unclear how to use them) over to sync groups. This simplifies the
pbxproj and makes it easier to develop the project from vs code or the
command line.
- PBXFileSystemSynchronizedRootGroup is a newer model type that matches
files in a directory automatically so you don't need to manually clean
up references in the pbxproj.

<img width="223" height="441" alt="Screenshot 2025-12-05 at 3 06 46 PM"
src="https://github.com/user-attachments/assets/2639307c-d347-4060-853a-72b6d96ec9a8"
/>


<!--
Please describe the motivation for this PR, and link to relevant GitHub
issues, forums posts, or feature requests.
-->

# How

- eyeball it.

# Test Plan

- App still builds.