[e2e] Update android emulator installation instructions in README.md (#41484)
# Why

In the instructions we provide a command to create the avd, but when the
system image is missing it fails.
Also the proposed avd is `x86_64`, our typical development devices are
ARM macbooks.

# How

Add the installation command so that the setup is more straightforward,
change the architecture to `arm64-v8a`

# Test Plan

Tested when setting up e2e on my machine