[docs] Adds release statuses (#41500)
# Why

We often label new features or APIs with a variety of different statuses
that mean many different things, like:
- Experimental
- Preview
- Alpha
- Beta

Sometimes these mean that something is new and we don't have enough
validation yet to feel confident recommending them. Sometimes it means
that we know of significant or minor issues. Sometimes we leave labels
on features permanently because we're not sure when to promote a feature
to the next release status.

This is confusing for developers and customers, because you have to ask
us directly what we mean by any of these terms.

This PR creates release statuses and defines them in a doc. The release
statuses are:
- Alpha (Could be "experimental" or "early access")
- Beta (Could be "preview")
- Stable
- Deprecated

This PR also links the "alpha" state to the Maestro test jobs
documentation, so users can see exactly what that means. The expectation
would be that going forward, any feature or API would link to this doc
to better describe where it is in our release sequence.

# Test Plan

Check out /more/release-statuses for the definition doc. Then check out
the pre-packaged jobs doc (the maestro job section) for a link to the
preview header.

<img width="1718" height="1440" alt="Screenshot 2025-12-08 at 12 57
47 PM"
src="https://github.com/user-attachments/assets/29d8f59c-5f43-4b7e-add0-cebbf7080844"
/>
<img width="1718" height="1440" alt="Screenshot 2025-12-08 at 12 57
37 PM"
src="https://github.com/user-attachments/assets/fdcfbc5c-635c-448b-9158-0578180bd8d4"
/>

---------

Co-authored-by: Aman Mittal <amandeepmittal@live.com>