[video][android] Set property values on calling thread. (#41533)
# Why

For most player properties, we have to schedule the ExoPlayer setter on
the main thread, since it has to be modified on the main thread.

So far we have also been scheduling our property setter on the main
thread for the sake of simplicity.
This introduces an issue though. Doing the following:
```
// Default player volume is 1
player.volume = 0.5
print(player.volume) // Will usually print 1
```

In normal use cases this didn't cause any issues, but it does in
property testers in `test-suite`, since they usually do the pattern
mentioned above.

# How

Set our property value on the calling thread, and schedule the ExoPlayer
property setting on the main thread.

# Test Plan

Tested in BareExpo on Android