[autolinking][android] Add `expo-max-sdk-override-plugin` gradle plugin (#40973)
# Why

Adds `expo-max-sdk-override-plugin` gradle plugin to change the default
manifest merger behaviour for `android:maxSdkVersion` conflicts.

This is meant to resolve following issues in the current expo sdk:

- expo-screen-capture - Requires `READ_MEDIA_IMAGES` ONLY for sdk 33 -
that is because for other SDK versions it uses methods that don't use
this permission, it defines the permissions with
`android:minSdkVersion="33" android:maxSdkVersion="33"`.
- expo-media-library - Requires `READ_MEDIA_IMAGES` without any limit
(through a config plugin)

The merged AndroidManifest.xml will have the following permission:
` <uses-permission android:name="android.permission.READ_MEDIA_IMAGES"
android:maxSdkVersion="33"/>`
The `android:minSdkVersion` works as we want, but
`android:maxSdkVersion` doesn't (we'd like it to disappear as well).

An obvious solution is to declare the premission in `expo-media-library`
with `tools-remove:maxSdkVersion`, but most people wouldn't think of it
when declaring permissions, this is especially true for third-party expo
libraries.

Another solution is to declare the permission without
`android:maxSdkVersion="33"` inside `expo-screen-capture`, but then for
projects that use just `expo-screen-capture` the store review process
may be more difficult, since they don't actually need the permission for
other api levels.

A third solution would be adding a config plugin parameter to
`expo-screen-capture` that can add `android:maxSdkVersion` to the
permission when true, but i feel like it may be a bit hard to explain
and would add unnecessary complexity to the usage.

I think that for our usages the best option is to override the default
merger behaviour. I couldn't come up with a situation where we'd like to
keep the default `android:maxSdkVersion` merging behaviour in this very
specific case. It should also keep us from introducing weird permission
interferences between libraries in the future (which is probably the
strongest argument for introducing these changes)

# How

1. Reads the manifest merge report
2. Based on the merge reports extracts all permissions that were defined
in multiple manifests and contain the `android:maxSdkVersion`. Also
saves the paths to `AndroidManifest.xml` in question.

Now we wan't to know if the `android:maxSdkVersion` is present in all
permission declarations or if there exists at least one permission
declaration that doesn't require it. We can't extract this info from the
blame report.

4. We parse manifests under the afformentioned paths and inside them
find permissions that are declared with `android:maxSdkVersion` in one
manifest and without it in another
5. We modify the final merged manifest and remove
`android:maxSdkVersion` from conflicting places.

I'm not sure if there is a better way to extract info about the
permissions and any info is hard to find because of lacking AGP
documentation. Let me know if you have any less tricky ideas!

Example output from the plugin:

```
---------- Expo Max Sdk Override Plugin ----------
>>> WARNING: Found 2 permission(s) with conflicting 'android:maxSdkVersion' declarations.
    - android.permission.READ_EXTERNAL_STORAGE
      > Defined WITH `android:maxSdkVersion` in: /Users/wojciechdrozdz/work/expo/packages/expo-image/android/build/intermediates/merged_manifest/debug/processDebugManifest/AndroidManifest.xml
      > Defined WITHOUT `android:maxSdkVersion` in: /Users/wojciechdrozdz/work/expo/apps/bare-expo/android/app/src/main/AndroidManifest.xml, /Users/wojciechdrozdz/work/expo/packages/expo-image-picker/android/build/intermediates/merged_manifest/debug/processDebugManifest/AndroidManifest.xml, /Users/wojciechdrozdz/work/expo/packages/expo-screen-capture/android/build/intermediates/merged_manifest/debug/processDebugManifest/AndroidManifest.xml, /Users/wojciechdrozdz/work/expo/packages/expo-file-system/android/build/intermediates/merged_manifest/debug/processDebugManifest/AndroidManifest.xml, /Users/wojciechdrozdz/work/expo/packages/expo-media-library/android/build/intermediates/merged_manifest/debug/processDebugManifest/AndroidManifest.xml
    - android.permission.READ_MEDIA_IMAGES
      > Defined WITH `android:maxSdkVersion` in: /Users/wojciechdrozdz/work/expo/packages/expo-screen-capture/android/build/intermediates/merged_manifest/debug/processDebugManifest/AndroidManifest.xml
      > Defined WITHOUT `android:maxSdkVersion` in: /Users/wojciechdrozdz/work/expo/packages/expo-media-library/android/build/intermediates/merged_manifest/debug/processDebugManifest/AndroidManifest.xml
>>> Removing 'android:maxSdkVersion' from these permissions in the final manifest to prevent runtime issues.
>>> Removed 'android:maxSdkVersion' from 2 instance(s) in the final manifest.
--------------------------------------------------
```


# Test Plan

Tested in BareExpo on Android, this removes the conflicts between
`expo-screen-capture` and `expo-media-library` while also highlighting a
similar issue with `READ_EXTERNAL_STORAGE`. Although in the case of
`READ_EXTERNAL_STORAGE` we don't really care since it does nothing since
api level 32 anyways.