[video] Add `averageBitrate` and `peakBitrate` for video tracks (#41532)
# Why

Currently we provide the `bitrate` field, which behaves the same as the
`bitrate` field in `ExoPlayer` - it reports peak bitrate or average if
peak is unavailable.

There are currently a few issues with this: 
1. There is no way to know which of those is reported.
2. On iOS we weren't able to read peak bitrate for sources in MPEG-4
container at all since the AVKit doesn't provide this information
3. I decided that i ***really*** don't like this ambiguous way of
reporting bitrate

# How

Add `averageBitrate` and `peakBitrate` fields for much clearer bitrate
reporting.

On iOS use the metadata fields to find the `btrt` box, which contains
bitrate information. This is not the most reliable solution, but should
work for most common sources where the source actually reports the peak
bitrate (for our test sources we have parity with ExoPlayer peak bitrate
reporting).

I also marked the bitrate field as deprecated

# Test Plan

Tested in BareExpo and Android