[video] Add test-suite tests for Android and iOS (#41455)
# Why

Expo-video doesn't have tests in test-suite

# How

Added tests that test pretty much all properties and events of the
player. I'm planning to cover `VideoView` functionality using e2e tests,
since most of them aren't possible to test with jest.

I didn't add the tests for web. For web we have to mound the video
player in a `VideoView` using the portal, but the data doesn't seem to
be loading in the portal making it impossible to test.

# Test Plan

Tested in test-suite on an iPhone 13 and Pixel 8